update starlink-grpc-go protos, fix build.
The project also fails to build with the current `go.mod`/`go.sum`
version for this dependency (see the associated issue[0]). This can be
side-stepped by pointing at the right commit of
`starlink-community/starlink-grpc-go`, however it introduces a new
challenge:

Running a router w/ `2022.01.0.mr8624-prod` and a dish w/
`992cafb5-61c7-46a3-9ef7-5907c8cf90fd.uterm.release` the
`starlink-community/starlink-grpc-go` generated protobuf types result in
incomplete output with "unknown bytes".

This PR fixes the build and the incomplete information by temporarily
using a fork of `starlink-grpc-go` with updated generated protos. Once
the upstream project has merged the updated protos the `replace`
directive to use the fork can be removed.

With this branch in place I was able to fetch complete information for
the software versions on my dish/router.

[0]: https://github.com/starlink-community/starlink-cli/issues/1